src/database.py
import os
import sys
import psycopg2
from psycopg2 import pool
from dotenv import load_dotenv
from typing import Optional

# Load environment variables at the module level
load_dotenv()

# --- Database Connection Pool ---
# We use a connection pool to manage PostgreSQL connections efficiently.
# The pool is initialized once when the module is loaded.
db_url = os.getenv("DB_URL")
connection_pool = None

def initialize_connection_pool():
    """
    Initializes the PostgreSQL connection pool.
    This should be called once when the application starts up.
    """
    global connection_pool
    if not db_url:
        print("Error: DB_URL environment variable is not set.", file=sys.stderr)
        sys.exit(1)

    try:
        print("Initializing database connection pool...")
        connection_pool = psycopg2.pool.SimpleConnectionPool(
            minconn=1,
            maxconn=10,
            dsn=db_url
        )
        print("Database connection pool initialized successfully.")
    except psycopg2.OperationalError as e:
        print(f"Error: Could not connect to the database and initialize pool.", file=sys.stderr)
        print(f"Please check your DB_URL and ensure the database server is running.", file=sys.stderr)
        print(f"Details: {e}", file=sys.stderr)
        sys.exit(1)

def get_db_connection():
    """
    Gets a connection from the pool.

    Returns:
        A psycopg2 connection object.

    Raises:
        Exception: If the connection pool is not initialized.
    """
    # TODO: Implement a more robust mechanism to ensure the pool is initialized,
    # perhaps in the FastAPI startup event.
    if connection_pool is None:
        raise Exception("Connection pool has not been initialized. Call initialize_connection_pool() first.")

    return connection_pool.getconn()

def release_db_connection(conn):
    """
    Releases a connection back to the pool.

    Args:
        conn: The psycopg2 connection object to release.
    """
    if connection_pool:
        connection_pool.putconn(conn)

# --- Caching Logic ---

def get_cached_html(file_hash: str) -> Optional[str]:
    """
    Retrieves cached HTML content from the database for a given file hash.

    Args:
        file_hash: The SHA-256 hash of the file.

    Returns:
        The cached HTML content as a string, or None if not found.
    """
    # TODO: Implement the database query logic.
    # This function will connect to the DB, query the html_cache table,
    # and return the html_content if a matching hash is found.
    print(f"TODO: Checking cache for hash: {file_hash}")

    conn = get_db_connection()
    # Placeholder logic
    release_db_connection(conn)

    return None

def cache_html(file_hash: str, html_content: str) -> None:
    """
    Saves new HTML content to the database cache.

    Args:
        file_hash: The SHA-256 hash of the file.
        html_content: The HTML content to cache.
    """
    # TODO: Implement the database insertion logic.
    # This function will connect to the DB and insert a new record
    # into the html_cache table.
    print(f"TODO: Caching HTML for hash: {file_hash}")

    conn = get_db_connection()
    # Placeholder logic
    release_db_connection(conn)

    pass

# Initialize the pool when the module is loaded.
# In a real FastAPI app, this would be better handled in a startup event.
initialize_connection_pool()
